requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
torch>=1.9.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.0
joblib>=1.0.0
xgboost>=1.4.0

Flask==3.1.2
flask-cors==6.0.2
Flask-SocketIO==5.6.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
eventlet==0.40.4
bcrypt==5.0.0
email-validator==2.3.0
torch>=2.0.0
torchvision>=0.15.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.3.0
joblib==1.4.2
xgboost==2.1.4
shap>=0.42.0
python-dotenv>=1.0.0
requests==2.32.4
openai==2.16.0
google-generativeai==0.8.6
SQLAlchemy==2.0.46
matplotlib==3.7.5
Pillow>=10.0.0
scipy>=1.9.0